changed news for fwa
