+++ afc_site/settings.py
@@ -32,7 +32,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'news',
+    'fwa',
 
     'django.contrib.admin',
     'django.contrib.auth',
