+++ fwa/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class NewsConfig(AppConfig):
+class FwaConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'news'
+    name = 'fwa'
