+++ fwa/models.py
@@ -1,12 +1,15 @@
 from django.db import models
 
 
+class Teams(models.Model):
+    name = models.CharField(max_length=50, unique=True)
+
+
 class News(models.Model):
-    # TODO: change date type from str to datetime
     date = models.DateTimeField()
-    team = models.CharField(max_length=50)
+    team = models.ForeignKey(Teams, on_delete=models.CASCADE)
     title = models.TextField()
-    source = models.URLField()
+    source = models.URLField(unique=True)
 
     def __str__(self):
         return f'<News: {self.title} from {self.source}>'
